+++ src/object/alga.py
@@ -1,4 +1,4 @@
-from src.ai.ai import AlgaAI
+from src.ai.alga_ai import AlgaAI
 from src.constants import ALGA_SURFACING_STEPS
 from src.object.pond_object import PondObject
 from src.object_kind import ObjectKind
