+++ src/object/alga_maker.py
@@ -1,6 +1,6 @@
 from random import randint
 
-from src.ai.ai import AlgaMakerAI
+from src.ai.alga_maker_ai import AlgaMakerAI
 from src.constants import MAX_ALGAE_TO_CREATE, MIN_ALGAE_TO_CREATE, ALGA_ENERGY_VALUE
 from src.object.alga import Alga
 from src.object.pond_object import PondObject
