+++ src/object/worm.py
@@ -1,4 +1,4 @@
-from src.ai.ai import WormAI
+from src.ai.worm_ai import WormAI
 from src.constants import WORM_FALLING_STEPS, WORM_BOUNCE_STEPS
 from src.object.pond_object import PondObject
 from src.object_kind import ObjectKind
